% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FEMr.objects.R
\name{FEM}
\alias{FEM}
\title{Define a Functional Data through a Finite Element basis}
\usage{
FEM(coefmat,basisobj)
}
\arguments{
\item{coefmat}{A vector or a matrix containg the coefficients of a Finite Element representation. The number of rows (or the vector's length) corresponds to the number of functions defining the basis, i.e.
the number of elements in the mesh. The number of columns corresponds to the number of functional replicates.}

\item{basisobj}{An object defining the Finite Element basis. It can be created with \link{create.FEM.basis}.}
}
\value{
An object of the class \code{FEM}. This contains a list with components \code{coefmat} and \code{basis}.
}
\description{
This is the constructur function for objects of the class FEM. Each function of FEMr that needs to create a
new Functional Object defined respect to the Finite Element (FE) basis, must call this function. Usually users do not need to call this function directly.
}
\examples{
library(FEMr)
data("mesh.2D.rectangular")
plot(mesh.2D.rectangular)
## FEM object with 1st order Finite Element basis
basisobj = create.FEM.basis(mesh.2D.rectangular, 1)
coeff <- sin(mesh.2D.rectangular$nodes[,1])*cos(mesh.2D.rectangular$nodes[,2])
FEM_object<- FEM(coeff, basisobj)
plot(FEM_object)
}

