% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{horseshoe2D}
\alias{horseshoe2D}
\title{Horseshoe domain}
\usage{
data(horseshoe2D)
}
\description{
The boundary and interior nodes and connectivity matrix of a triangular mesh of the horseshoe domain. 
This dataset can be used to create a \code{MESH.2D} object with the function \code{create.MESH.2D}. 
The variables are:
\itemize{
   \item boundary_nodes. The nodes in the boundary.
   \item boundary_segments. The ssegments of the boundary.
   \item locations. The interior nodes of the mesh. 
}
}
