% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{hub2.5D}
\alias{hub2.5D}
\title{Hub domain}
\usage{
data(hub2.5D)
}
\description{
The nodes and connectivity matrix of a triangular mesh of a manifold representing a hub geometry. 
This dataset can be used to create a \code{MESH.2.5D} object with the function \code{create.MESH.2.5D}.
The variables are:
\itemize{
   \item hub2.5D.nodes. The nodes of the mesh.
   \item hub2.5D.triangles. The triangles of the mesh.
}
}
