% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FEMr.objects.R
\name{image.FEM}
\alias{image.FEM}
\title{Draw a Image Plot of a Functional Object of the class FEM}
\usage{
\method{image}{FEM}(x, num_refinements, ...)
}
\arguments{
\item{x}{A functional object of the class FEM.}

\item{num_refinements}{An natural number specifying how many bisections should by applied to each triangular element for
plotting purpose. This functionality is useful where a discretization with 2nd order Finite Element has been applied.}

\item{...}{Arguments representing the graphical parameters to be passed to methods, see \link[rgl]{plot3d}.}
}
\description{
Functional objects created with the function \code{FEM} or returned by \code{smooth.FEM.basis}, \code{smooth.FEM.PDE.basis} or
\code{smooth.FEM.PDE.SV.basis} can be visualized trough an image plot.
}
\examples{
data(MeuseData)
data(MeuseBorder)
order=1
mesh <- create.MESH.2D(nodes = MeuseData[,c(2,3)], segments = MeuseBorder, order = order)
data = log(MeuseData[,7])
basisobj = create.FEM.basis(mesh, order)
lambda = 10^3.5
ZincMeuse = smooth.FEM.basis(observations = data, basisobj = basisobj, lambda = lambda)
## plot of the FEM object representing the fitted function
image(ZincMeuse$fit.FEM)
## plot the FEM object representing the misfit
image(ZincMeuse$PDEmisfit.FEM)
}

