% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{MeuseData}
\alias{MeuseData}
\title{Meuse river data set}
\format{A data frame with 155 rows and 12 variables}
\source{
\url{https://cran.r-project.org/web/packages/gstat/gstat.pdf}
}
\description{
This data set gives locations and top soil heavy metal concentrations (ppm). It is used in many examples to illustrate FEMr features. 
For further details on this dataset, see \code{meuse.all} in \url{https://cran.r-project.org/web/packages/gstat/gstat.pdf}
}
\details{
\itemize{
  \item x. numeric vector indicating the x-coordinate.
  \item y. numeric vector indicating the y-coordinate.
  \item zinc. topsoil zinc concentration, ppm.
  ...
}
}

