% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mesh.2D.R
\name{plot.TRIMESH2D}
\alias{plot.TRIMESH2D}
\title{Plot a TRIMESH2D object}
\usage{
\method{plot}{TRIMESH2D}(x, ...)
}
\arguments{
\item{x}{A TRIMESH2D object defining the mesh, as generated by \code{create.Mesh.2D} or \code{refine.Mesh.2D}.}

\item{...}{Arguments representing graphical options to be passed to \link[graphics]{par}.}
}
\description{
Plot a mesh TRIMESH2D object, generated by \code{create.MESH.2D} or \code{refine.MESH.2D}. Circles indicate the mesh nodes.
}
\examples{
Upload the Meuse data and a domain boundary
data(MeuseData)
data(MeuseBorder)
Create a triangular mesh with the provided boundary
mesh <- create.MESH.2D(nodes = MeuseData[,c(2,3)], segments = MeuseBorder, order = 1)
plot(mesh)
}

