% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{quasicircle2D}
\alias{quasicircle2D}
\title{Quasicircle2D domain}
\usage{
data(quasicircle2D)
}
\description{
The boundary and interior nodes and connectivity matrix of a triangular mesh of a quasicircular domain, together 
with a non-stationary field observed over the nodes of the mesh.  
This dataset can be used to create a \code{MESH.2D} object with the function \code{create.MESH.2D} and to test
the smooth.FEM function.
The variables are:
\itemize{
   \item boundary_nodes. The nodes in the boundary.
   \item boundary_segments. The ssegments of the boundary.
   \item locations. The interior nodes of the mesh. 
   \item data. The vector of observations. 
}
}
