% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FEMr.smoothing_R.R
\name{R_elementProperties}
\alias{R_elementProperties}
\title{Compute some properties for each triangular element of the mesh}
\usage{
R_elementProperties(mesh)
}
\arguments{
\item{mesh}{A  MESH2D mesh object representing the triangular mesh. This can be created with  \code{\link{create.MESH.2D}}.}
}
\value{
A list with the following variables:
\item{\code{detJ}}{The determinant of the transformation from the reference triangle to the nodes of the i-th triangle. It's values is also the double of the area of each triangle of the basis.}
\item{\code{transf}}{A matrix such that \code{transf[i,,]} is the 2-by-2 tranformation matrix that transforms the nodes of the reference triangle to the nodes of the i-th triangle.}
\item{\code{metric}}{A matrix such that \code{metric[i,,]} is the 2-by-2 matrix \code{transf[i,,]^{-1}*transf[i,,]^{-T}}. This matrix is usuful for the computation
of the integrals over the elements of the mesh.}
}
\description{
Only executed when the function \code{create.FEM.basis} is run with the option \code{CPP_CODE} = \code{FALSE}. For each linear map that transforms the ith triangle in the reference element, three properties are computed.
These are used for the computation of the integrals necessary to build the mass and stiffness matrix.
}

