% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FEMr.objects.R
\name{FEM}
\alias{FEM}
\title{Define a function by a Finite Element basis expansion}
\usage{
FEM(coefmat,basisobj)
}
\arguments{
\item{coefmat}{A vector or a matrix containg the coefficients for the Finite Element basis expansion. The number of rows (or the vector's length) corresponds to the number of basis in code{basisobj}. 
The number of columns corresponds to the number of functional replicates.}

\item{basisobj}{An object defining the Finite Element basis, created by \link{create.FEM.basis}.}
}
\value{
An object of the class \code{FEM}. This contains a list with components \code{coefmat} and \code{basisobj}.
}
\description{
This function defines a FEM object. It is not called directly by users.
}
\examples{
## Upload a triangular mesh and plot it
data("mesh.2D.rectangular")
plot(mesh.2D.rectangular)
## Create a linear Finite Element basis
basisobj = create.FEM.basis(mesh.2D.rectangular, 1)
## Define a sinusoidal function as expansion of these basis and plot it
coeff <- sin(mesh.2D.rectangular$nodes[,1])*cos(mesh.2D.rectangular$nodes[,2])
FEM_object<- FEM(coeff, basisobj)
plot(FEM_object)
}

