% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InferenceDataObject.R
\docType{class}
\name{inferenceDataObject-class}
\alias{inferenceDataObject-class}
\alias{inferenceDataObject}
\title{Class for inference data}
\description{
A class that contains all possible information for inference over linear parameters in spatial regression with
differential regularization problem. This object can be used as parameter in smoothing function of the fdaPDE library [smooth.FEM].
}
\details{
#Warning
At least one between test and interval must be nonzero. \code{dim}, \code{coeff} and \code{beta0} need to be coherent.
The usage of [inferenceDataObjectBuilder] is recommended for the construction of an object of this class.
}
\section{Slots}{

\describe{
\item{\code{test}}{A vector of integers taking value 0, 1 or 2; in the first case no test is performed, in the second one-at-the-time tests are performed,
in the third a simultaneous test is performed.}

\item{\code{interval}}{A vector of integers taking value 0, 1, 2 or 3; In the first case no confidence interval is computed, in the second case one-at-the-time confidence intervals are computed, 
in the third case simultaneous confidence intervals are computed, in the fourth case Bonferroni confidence intervals are computed.}

\item{\code{type}}{A vector of integers taking value 1, 2 or 3, corresponding to Wald, speckman or eigen-sign-flip implementation 
of the inference analysis.}

\item{\code{exact}}{An integer taking value 1 or 2. If 1 an exact computation of the test statistics will be performed,
whereas if 2 an approximated computation will be carried out.}

\item{\code{dim}}{Number of covariates taken into account in the linear part of the regression problem.}

\item{\code{coeff}}{A matrix of numeric coefficients with columns of dimension \code{dim} and each row represents a linear combination of the linear parameters to be tested and/or to be
estimated via confidence interval.}

\item{\code{beta0}}{Vector of null hypothesis values for the linear parameters of the model. Used only if \code{test} is not 0.}

\item{\code{quantile}}{Quantile needed for confidence intervals. Used only if interval is not 0.}

\item{\code{n_flip}}{An integer representing the number of permutations in the case of eigen-sign-flip test.}

\item{\code{tol_fspai}}{A real number greater than 0 specifying the tolerance for FSPAI algorithm, in case of non-exact inference.}

\item{\code{definition}}{An integer taking value 0 or 1. If set to 1, the class will be considered as created by the function [inferenceDataObjectBuilder],
leading to avoid some of the checks that are performed on inference data within smoothing functions.}
}}

