% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InferenceDataObject.R
\name{inferenceDataObjectBuilder}
\alias{inferenceDataObjectBuilder}
\title{Constructor for inferenceDataObject class}
\usage{
inferenceDataObjectBuilder<-function(test = NULL, 
interval = NULL, 
type = "wald", 
exact = "False", 
dim = NULL, 
coeff = NULL, 
beta0 = NULL, 
level = 0.05,
definition=1)
}
\arguments{
\item{test}{A string defining the type of test to be performed. The default is NULL, and can take values 'one-at-the-time' or 'simultaneous'.
If the value is NULL, no test is performed, and the \code{interval} parameter need to be not NULL. If it takes value
'one-at-the-time', the parameters \code{beta0}, \code{dim}, \code{coeff} are taken into account, and one-at-the-time tests will be performed. If it takes value 
'simultaneous', a global simultaneous test will be performed.}

\item{interval}{A string defining the type of confidence intervals to be computed. The default is NULL, and can take value 'one-at-the-time' 'simultaneous' and 'bonferroni'.
If the value is NULL, no interval will be computed, and the \code{test} parameter needs to be set. Otherwise one at the time, simultaneous or Bonferroni correction intervals will be computed.
If it is not NULL, the parameter \code{level} will be taken into account. Up to now, confidence intervals can be computed only in the Wald implementation.}

\item{type}{A string defining the type of implementation for the inferential analysis. The possible values are three:
'wald'(default), 'sandwich' or 'permutational', corresponding to the three possible methods developed in Ferraccioli....}

\item{exact}{A string used to decide the method used to estimate the statistics variance.
The possible values are: 'True' and 'False'(default). In the first case the evaluation is exact but computationally very expensive.
In the second case an approximate method is used, leading to a lower accuracy, but faster computation.}

\item{dim}{Number of the covariates, defaulted to NULL.}

\item{coeff}{A matrix, with \code{dim} number of columns, of numeric coefficients, defaulted to NULL. If this parameter is NULL, it will be defaulted to an identity matrix.}

\item{beta0}{Vector of real numbers (default NULL). It is used only if the \code{test} parameter is set, and has length the number of rows of matrix \code{coeff}. If \code{test} is set and \code{beta0} is NULL,
will be set to a vector of zeros.}

\item{level}{Level of significance, defaulted to 0.05. It is taken into account only if \code{intrval} is set.}

\item{definition}{A 0-1 integer. If it is set, and the procedure outcomes a valid inferenceDataObject, some checks will be avoided inside [smooth.FEM].}
}
\value{
The output is a well defined [inferenceDataObject], that can be used as parameter in the [smooth.FEM] function.
}
\description{
A function that build an [inferenceDataObject]. In the process of construction many checks over the input parameters are carried out so that the output is a well defined object,
that can be used as parameter in [smooth.FEM] function. Notice that this constructor ensures well-posedness of the object, but a further check on consistency with smooth.FEM parameters will be carried out inside that function.
}
\examples{
obj1<-inferenceDataObjectBuilder(test = "simultaneous", interval = NULL, dim = 4);
obj2<-inferenceDataObjectBuilder(interval = "one-at-the-time", dim = 5, level = 0.01);
}
