% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{MeuseBorder}
\alias{MeuseBorder}
\title{Meuse Border description data set}
\format{A data frame with 52 rows and 2 variables}
\description{
This file describes the boundaries of the domain of the Meuse dataset.
}
\details{
\itemize{
  \item V1. An integer vector with the indices of the locations where the segments defining the border start from.
  \item V2. An integer vector with the indices of the locations where the segments defining the border end to
}
}

