\name{MeuseBorder}
\alias{MeuseBorder}
\title{Border of Meuse river data set}
\description{
This data set gives the descriptions of the domain's border of the Meuse river data set
}
\format{
  This data frame contains the following columns:
  \describe{
    \item{V1}{Index of the starting point of the border segments}
    \item{V2}{Index of the ending point of the border segments}
  } 
} 
\usage{
data(MeuseBorder)
}

\keyword{datasets}
\examples{
data(MeuseBorder)
summary(MeuseBorder)
}
