% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\name{MeuseBorder}
\alias{MeuseBorder}
\title{Boundary of the Meuse River data set}
\format{A data frame with 52 rows and 2 variables.}
\description{
This file provides the boundary of the domain of the Meuse dataset  {\link{MeuseData}}.
}
\details{
\itemize{
  \item V1. A vector having as entries the indices of the locations in {\link{MeuseData}} where a boundary segment starts from.
  \item V2. A vector having as entries the indices of the locations in {\link{MeuseData}} where a boundary segment ends to
}
}

