% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mesh.2D.R
\name{plot.TRIMESH2D}
\alias{plot.TRIMESH2D}
\title{Plot a mesh object of the class TRIMESH2D}
\usage{
\method{plot}{TRIMESH2D}(x, ...)
}
\arguments{
\item{x}{An object of the class TRIMESH2D defining the mesh. It can be created with \code{create.Mesh.2D} or \code{refine.Mesh.2D}.}

\item{...}{Arguments representing the graphical parameters to be passed to methods, see \link[graphics]{par}.}
}
\description{
Mesh objects created with the functions \code{create.MESH.2D} or \code{refine.MESH.2D} can be plotted.
}
\examples{
library(FEMr)

data(MeuseData)
data(MeuseBorder)
mesh <- create.MESH.2D(nodes = MeuseData[,c(2,3)], segments = MeuseBorder, order = 1)
plot(mesh)
}

