% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/FEMr.smoothing_R.R
\name{R_elementProperties}
\alias{R_elementProperties}
\title{Compute some properties for each triangular element of the mesh}
\usage{
R_elementProperties(mesh)
}
\arguments{
\item{mesh}{A mesh object of the class TRIMESH2D. This can be created with  \code{\link{create.MESH.2D}}.}
}
\value{
A list with the following variables:
\item{\code{J}}{The area of each triangle of the basis.}
\item{\code{transf}}{A matrix such that \code{transf[i,,]} is the 2-by-2 tranformation matrix that transforms the nodes of the reference triangle to the nodes of the i-th triangle.}
\item{\code{metric}}{A matrix such that \code{metric[i,,]} is the 2-by-2 matrix \code{transf[i,,]^{-1}*transf[i,,]^{-T}}. This matrix is usuful for the computation
of the integrals over the elements of the mesh.}
}
\description{
For each linear map that transforms the ith triangle in the reference element, three properties are computed.
These are used for the computation of the integrals necessary to compute the mass and stiffness matrix.
This version of the function is implemented using only R code. It is called by \code{create.FEM.basis} when \code{CPP_CODE} is \code{FALSE},
in fact this quantities are used only when the folloing computations are executed with the flag \code{CPP_CODE = TRUE}.
}
\references{
Sangalli, L.M., Ramsay, J.O. & Ramsay, T.O., 2013. Spatial spline regression models. Journal of the Royal Statistical Society. Series B: Statistical Methodology, 75(4), pp.681.703.
}

