% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FEMr.smoothing.R
\name{smooth.FEM.basis}
\alias{smooth.FEM.basis}
\title{Compute a spatial regression model with differential regularization: stationary and isotropic case}
\usage{
smooth.FEM.basis(locations = NULL, observations, basisobj, lambda, 
       covariates = NULL, BC = NULL, GCV = FALSE, CPP_CODE = TRUE)
}
\arguments{
\item{locations}{A #observations-by-2 matrix where each row specifies the spatial coordinates of the corresponding observation in \code{observations}. ***}

\item{observations}{A #observations vector with the observed values on the domain. 
The locations of the observations can be specified with the \code{locations} 
argument, otherwise the locations are intented to be the corresponding nodes of the mesh****. 
\code{NA} values are admissible to indicate that the node is not associated with any observed data value.}

\item{basisobj}{A FEM object describing the Finite Element basis, as created by \code{\link{create.FEM.basis}}.}

\item{lambda}{A scalar or vector of smoothing parameters.}

\item{covariates}{A #observations-by-#covariates matrix where each row represents the covariates associated with the corresponding observed data value in \code{observations}.}

\item{BC}{A list with two vectors: 
\code{BC_indices}, a vector with the indices in \code{nodes} of boundary nodes where a Dirichlet Boundary Condition should be applied;
\code{BC_values}, a vector with the values the spatial field must take at the nodes indicated in \code{BC_Indices}.}

\item{GCV}{Boolean. If \code{TRUE} the following quantities are computed: the trace of the smoothing matrix, the estimated error standard deviation,  and 
the Generalized Cross Validation criterion, corresponding the values of the smoothing parameter specified in \code{lambda}.}

\item{CPP_CODE}{Boolean. If \code{TRUE} it avoids the computation of some additional quantities, that are not necessary if the 
functions using the FEM basis are called with the flag \code{CPP_CODE=TRUE}}
}
\value{
A list with the following variables:
\item{\code{fit.FEM}}{A FEM object that represents the fitted spatial field.}
\item{\code{PDEmisfit.FEM}}{A FEM object that represents the Laplacian of the estimated spatial field.}
\item{\code{beta}}{If covariates is not \code{NULL}, a #covariates vector with the regression coefficients associated with each covariate.}
\item{\code{edf}}{If GCV is \code{TRUE}, a scalar or vector with the trace of the smoothing matrix for each value of the smoothing parameter specified in \code{lambda}.}
\item{\code{stderr}}{If GCV is \code{TRUE}, a scalar or vector with the estimate of the standard deviation of the error for each value of the smoothing parameter specified in \code{lambda}.}
\item{\code{GCV}}{If GCV is \code{TRUE}, a  scalar or vector with the value of the GCV criterion for each for each value of the smoothing parameter specified in \code{lambda}.}
}
\description{
This function implements a spatial regression model with differential regularization; isotropic and stationary case. In particular, the regularizing term involves the Laplacian of the spatial field. Space-varying covariates can be included in the model. The technique accurately handle data distributed over irregularly shaped domains. Moreover, various conditions can be imposed at the domain boundaries.
}
\examples{
## Upload the Meuse data and a domain boundary for these data
data(MeuseData)
data(MeuseBorder)
## Create a triangular mesh for these data with the provided boundary
order=1
mesh <- create.MESH.2D(nodes = MeuseData[,c(2,3)], segments = MeuseBorder, order = order)
plot(mesh)
data = log(MeuseData[,7])
basisobj = create.FEM.basis(mesh, order)
lambda = 10^3.5
ZincMeuse = smooth.FEM.basis(observations = data, basisobj = basisobj, lambda = lambda)
plot(ZincMeuse$fit.FEM)
}
\references{
Sangalli, L.M., Ramsay, J.O. & Ramsay, T.O., 2013. Spatial spline regression models. Journal of the Royal Statistical Society. Series B: Statistical Methodology, 75(4), pp. 681-703.
}

